feat: Estilizado Content Pt2
